break up alignment.cpp and alignment_emitter.cpp in anticipation of shipping parts to libvgio
