+++ src/alignment.hpp
@@ -14,13 +14,11 @@
 #include "htslib/sam.h"
 #include "htslib/vcf.h"
 #include "handle.hpp"
-#include "gafkluge.hpp"
+#include "alignment_io.hpp"
 
 namespace vg {
 
 const char* const BAM_DNA_LOOKUP = "=ACMGRSVTWYHKDBN";
-
-const uint64_t DEFAULT_PARALLEL_BATCHSIZE = 512;
 
 int hts_for_each(string& filename, function<void(Alignment&)> lambda);
 int hts_for_each_parallel(string& filename, function<void(Alignment&)> lambda);
@@ -55,29 +53,6 @@ size_t fastq_paired_two_files_for_each_parallel(const string& file1, const strin
 size_t fastq_paired_two_files_for_each_parallel_after_wait(const string& file1, const string& file2,
                                                            function<void(Alignment&, Alignment&)> lambda,
                                                            function<bool(void)> single_threaded_until_true);
-
-// single gaf
-bool get_next_alignment_from_gaf(const HandleGraph& graph, htsFile* fp, kstring_t& s_buffer, gafkluge::GafRecord& g_buffer,
-                                 Alignment& alignment);
-bool get_next_interleaved_alignment_pair_from_gaf(const HandleGraph& graph, htsFile* fp, kstring_t& s_buffer,
-                                                  gafkluge::GafRecord& g_buffer, Alignment& mate1, Alignment& mate2);
-size_t gaf_unpaired_for_each(const HandleGraph& graph, const string& filename, function<void(Alignment&)> lambda);
-size_t gaf_paired_interleaved_for_each(const HandleGraph& graph, const string& filename,
-                                       function<void(Alignment&, Alignment&)> lambda);
-// parallel gaf
-size_t gaf_unpaired_for_each_parallel(const HandleGraph& graph, const string& filename,
-                                      function<void(Alignment&)> lambda,
-                                      uint64_t batch_size = DEFAULT_PARALLEL_BATCHSIZE);
-size_t gaf_paired_interleaved_for_each_parallel(const HandleGraph& graph, const string& filename,
-                                                function<void(Alignment&, Alignment&)> lambda,
-                                                uint64_t batch_size = DEFAULT_PARALLEL_BATCHSIZE);
-size_t gaf_paired_interleaved_for_each_parallel_after_wait(const HandleGraph& graph, const string& filename,
-                                                           function<void(Alignment&, Alignment&)> lambda,
-                                                           function<bool(void)> single_threaded_until_true,
-                                                           uint64_t batch_size = DEFAULT_PARALLEL_BATCHSIZE);
-// gaf conversion
-gafkluge::GafRecord alignment_to_gaf(const HandleGraph& graph, const Alignment& aln, bool cs_cigar = true, bool base_quals = true);
-void gaf_to_alignment(const HandleGraph& graph, const gafkluge::GafRecord& gaf, Alignment& aln);
 
 bam_hdr_t* hts_file_header(string& filename, string& header);
 bam_hdr_t* hts_string_header(string& header,
@@ -193,12 +168,6 @@ pair<int32_t, int32_t> compute_template_lengths(const int64_t& pos1, const vecto
     const int64_t& pos2, const vector<pair<int, char>>& cigar2);
 
 int32_t sam_flag(const Alignment& alignment, bool on_reverse_strand, bool paired);
-short quality_char_to_short(char c);
-char quality_short_to_char(short i);
-string string_quality_char_to_short(const string& quality);
-string string_quality_short_to_char(const string& quality);
-void alignment_quality_char_to_short(Alignment& alignment);
-void alignment_quality_short_to_char(Alignment& alignment);
 void parse_rg_sample_map(char* hts_header, map<string, string>& rg_sample);
 int alignment_to_length(const Alignment& a);
 int alignment_from_length(const Alignment& a);
